+++ PRISTINE/PRISTINE_essentials.py
@@ -25,85 +25,17 @@ from shutil import copyfile
 import re
 import seaborn as sns
 import matplotlib.pyplot as plt
-from collections import Counter
-from dataclasses import dataclass, field
-from typing import Optional, Dict
-import time
-import yaml
 from pathlib import Path
 
-
-@dataclass
-class InputPaths:
-    raw_dir: Optional[str] = None
-    prokka_dir: Optional[str] = None
-    panaroo_dir: Optional[str] = None
-
-
-@dataclass
-class Primer3Params:
-    global_params: Dict[str, object] = field(default_factory=dict)
-    design_params: Dict[str, object] = field(default_factory=dict)
-
-
-@dataclass
-class SNPPrimerDesignParams:
-    snp_window_size: int
-    snp_top_n: int
-    min_snps: int
-
-
-@dataclass
-class ValidationConfig:
-    perform: str
-    database: str
-    pblat_min_identity: float
-    match_median_filter_tolerance: int
-
-
-@dataclass
-class Config:
-    input_type: str
-    input_paths: InputPaths
-    output_dir: str
-    max_cores: int
-    aligner: str
-    snp_avg_prop_threshold: float
-    primer3_config_file: Optional[str] = None
-    primer3: Primer3Params = field(default_factory=Primer3Params)
-    snp_primer_design: SNPPrimerDesignParams = field(default_factory=SNPPrimerDesignParams)
-    validation: ValidationConfig = field(default_factory=ValidationConfig)
-
-    def __post_init__(self):
-        allowed_inputs = {"raw", "prokka", "panaroo"}
-        if self.input_type not in allowed_inputs:
-            raise ValueError(f"`input_type` must be one of {allowed_inputs}, got: {self.input_type}")
-
-        active_dir = {
-            "raw": self.input_paths.raw_dir,
-            "prokka": self.input_paths.prokka_dir,
-            "panaroo": self.input_paths.panaroo_dir,
-        }[self.input_type]
-        if not active_dir:
-            raise ValueError(f"{self.input_type}_dir must be set in `input_paths` for input_type = {self.input_type}")
-
-        if self.primer3_config_file and self.primer3.global_params:
-            print("Warning: primer3_config_file is set. Inline primer3 parameters will be ignored.")
-
-
-class ConfigLoader:
-    @staticmethod
-    def load(path: str) -> Config:
-        with open(path, "r") as f:
-            raw = yaml.safe_load(f)
-
-        # Parse nested dataclasses manually
-        raw["input_paths"] = InputPaths(**raw.get("input_paths", {}))
-        raw["primer3"] = Primer3Params(**raw.get("primer3", {}))
-        raw["snp_primer_design"] = SNPPrimerDesignParams(**raw.get("snp_primer_design", {}))
-        raw["validation"] = ValidationConfig(**raw.get("validation", {}))
-
-        return Config(**raw)
+# Import configuration classes from config module
+from config import (
+    Config,
+    ConfigLoader,
+    InputPaths,
+    Primer3Params,
+    SNPPrimerDesignParams,
+    ValidationConfig,
+)
 
 
 class MSAStrategy:
@@ -2319,7 +2251,19 @@ class QuasiAlignmentStrategy():
   
         
 if __name__ == "__main__":
-    config = ConfigLoader.load("config.yaml")
+    parser = argparse.ArgumentParser(
+        description="PRISTINE: PRimer-based Identification Suite Targeting Informative Nucleotide Elements"
+    )
+    parser.add_argument(
+        "--config",
+        type=str,
+        default="config.yaml",
+        help="Path to the configuration YAML file (default: config.yaml in current directory)"
+    )
+    args = parser.parse_args()
+
+    # Load configuration from the specified path
+    config = ConfigLoader.load(args.config)
     runner = MSAStrategy(config)
     runner.run()
