+++ PRISTINE/toy_dataset/data/raw_genomes/Yersinia_pseudotuberculosis/Yersinia_pseudotuberculosis_1.fasta
@@ -1,4 +1,4 @@
->Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_accn|CP130901   Yersinia pseudotuberculosis strain SP-1303 chromosome, complete genome.   [Yersinia pseudotuberculosis SP-1303 | 633.777]
+>Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_accn|CP130901   Yersinia pseudotuberculosis strain SP-1303 chromosome, complete genome.   [Yersinia pseudotuberculosis SP-1303 | 633.777]
 gtgtcactttcgctttggcagcagtgtcttgcccgattgcaggatgagttacctgccaca
 gaatttagtatgtggatacgccccctacaggcggaactgagtgacaatactctggcgctt
 tacgcacctaatcgttttgtactggactgggtccgtgataagtacttaaacaatatcaat
