+++ PRISTINE/toy_dataset/data/raw_genomes/Yersinia_pseudotuberculosis/Yersinia_pseudotuberculosis_2.fasta
@@ -1,4 +1,4 @@
->Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_accn|LT596221   Yersinia pseudotuberculosis isolate NZYP4713 genome assembly, chromosome: 1.   [Yersinia pseudotuberculosis strain NZYP4713 | 633.88]
+>Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_Yersinia_pseudotuberculosis_accn|LT596221   Yersinia pseudotuberculosis isolate NZYP4713 genome assembly, chromosome: 1.   [Yersinia pseudotuberculosis strain NZYP4713 | 633.88]
 gtgtcactttcgctttggcagcagtgtcttgcccgattgcaggatgagttacctgccaca
 gaatttagtatgtggatacgccccctacaggcggaactgagtgacaatactctggcgctt
 tacgcacctaatcgttttgtactggactgggtccgtgataagtacttaaacaatatcaat
